\name{epade-package}
\alias{epade-package}
\alias{epade}
\docType{package}
\title{
Easy Plots
}
\description{
This package is a collection of statistical plots.
They are aimed at making fast overview plots from a data frame without elaborate preparations of data.
It is my first R package. The main motivation for making it was to learn R.
At the moment there is no error protection.
Be careful if you use any of the statistical tests in the plots, since they do not necessarily make sense.
}
\details{
\tabular{ll}{
Package: \tab epade\cr
Title: \tab Easy Plots\cr
Type: \tab Package\cr
Version: \tab 0.2.2\cr
Date: \tab 2011-08-15\cr
Depends: \tab plotrix, R (>= 2.12)\cr
Suggests: \tab survival\cr
License: \tab GPL version 2 or newer\cr
LazyLoad: \tab yes\cr
}
}
\author{
ADES\cr
Maintainer: <ades-s@web.de>
}
