\name{epade-package}
\alias{epade-package}
\alias{epade}
\docType{package}
\title{
Easy Plots
}
\description{
This package is a collection of statistical plots.
They are aimed at making fast overview plots from a data frame without elaborate preparations of data.
It is my first R package. The main motivation for making it was to learn R.
At the moment there is no error protection.
Be careful if you use any of the statistical tests in the plots, since they do not necessarily make sense.
}
\details{
\tabular{ll}{
Package: \tab epade\cr
Title: \tab Easy Plots\cr
Type: \tab Package\cr
Version: \tab 0.5.1\cr
Date: \tab 2022-10-25\cr
Depends: \tab plotrix, R (>= 2.12)\cr
Suggests: \tab survival\cr
License: \tab GPL (>= 2) \cr
LazyLoad: \tab yes\cr
}
}
\author{
Andreas Schulz\cr
Maintainer: <ades-s@web.de>
}
